# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PhoneMail API",
    "version": "0.1.0",
...
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/api/otp/request": { "post": { "requestBody": { "$ref": "#/components/requestBodies/OtpRequest" }, "responses": { "202": { "description": "Challenge created; code is delivered out of band. A local OTP sink, when configured in development, is not returned by this API.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OtpChallenge" } } } }, "400": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/api/otp/verify": { "post": { "requestBody": { "$ref": "#/components/requestBodies/OtpVerify" }, "responses": { "200": { "description": "Consumes only a generic supported proof; signup/login/phone-change must use their operation routes" }, "400": { "$ref": "#/components/responses/Error" } } } },
    "/api/otp/webhook": { "post": { "description": "Local mock provider event only; not a Twilio webhook. Uses X-Provider-Timestamp and X-Provider-Signature over the raw JSON body.", "parameters": [{ "$ref": "#/components/parameters/ProviderTimestamp" }, { "$ref": "#/components/parameters/ProviderSignature" }], "requestBody": { "$ref": "#/components/requestBodies/MockProviderEvent" }, "responses": { "200": { "description": "Accepted or duplicate event" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" } } } },
...
    "/api/search/messages": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 } }, { "name": "cursor", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Historical authorized search, filtered before pagination" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/api/drafts": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "responses": { "200": { "description": "Owner-only revisioned draft list; bodies are loaded on demand", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DraftPage" } } } } } }, "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "requestBody": { "$ref": "#/components/requestBodies/Draft" }, "responses": { "201": { "description": "Created private revisioned draft", "content": { "application/json": { "schema": { "type": "object", "required": ["draft"], "properties": { "draft": { "$ref": "#/components/schemas/Draft" } } } } } }, "400": { "$ref": "#/components/responses/Error" } } } },
    "/api/drafts/{id}": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/DraftId" }], "responses": { "200": { "description": "Owner-only draft with current revision" }, "404": { "$ref": "#/components/responses/Error" } } }, "patch": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/DraftId" }, { "$ref": "#/components/parameters/IfMatch" }], "requestBody": { "$ref": "#/components/requestBodies/Draft" }, "responses": { "200": { "description": "Updated draft with ETag revision" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }, "delete": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/DraftId" }, { "name": "revision", "in": "query", "schema": { "type": "integer", "minimum": 1 } }], "responses": { "204": { "description": "Deleted owner draft and emitted a sync tombstone" }, "409": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/api/drafts/{id}/send": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/DraftId" }, { "$ref": "#/components/parameters/IfMatchRequired" }, { "$ref": "#/components/parameters/IdempotencyKeyRequired" }], "requestBody": { "$ref": "#/components/requestBodies/DraftSend" }, "responses": { "201": { "description": "Atomically committed message, recipient copies, mailbox states, sync events and outbox job; retry with same key and payload" }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "428": { "$ref": "#/components/responses/Error" } } } },
    "/api/sync": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string", "pattern": "^(0|[1-9][0-9]*)$" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }], "responses": { "200": { "description": "Account-scoped incremental change page, including deletion tombstones", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncChangePage" } } } }, "410": { "$ref": "#/components/responses/Error" } } } },
    "/api/sync/snapshot": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "name": "snapshotId", "in": "query", "schema": { "type": "string", "format": "uuid" } }, { "name": "cursor", "in": "query", "schema": { "type": "string", "description": "Opaque keyset cursor for large snapshots; 0 restarts paging" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }], "responses": { "200": { "description": "Consistent current-state pages; snapshots above 10,000 entities use bounded keyset streaming. Message and draft bodies are omitted; incrementalCursor is available on the final page.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncSnapshotPage" } } } }, "404": { "$ref": "#/components/responses/Error" }, "410": { "$ref": "#/components/responses/Error" } } } },
    "/api/sync/snapshot/{id}/close": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "204": { "description": "Releases the caller-owned snapshot and any pinned read transaction" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
    "/api/users/{id}/profile-picture": { "get": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Authenticated profile picture subject to profile visibility" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
    "/api/mail/compose": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyRequired" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["to", "body"], "additionalProperties": false, "properties": { "to": { "type": "array", "minItems": 1, "maxItems": 50, "items": { "type": "string" } }, "cc": { "type": "array", "maxItems": 50, "items": { "type": "string" } }, "subject": { "type": "string", "maxLength": 200 }, "body": { "type": "string", "minLength": 1, "maxLength": 100000 }, "attachmentIds": { "type": "array", "maxItems": 10, "items": { "type": "string", "format": "uuid" } } } } } } }, "responses": { "201": { "description": "Message committed locally; internal recipients are delivered locally and external To/CC recipients are queued for SMTP. Relay acceptance is not inbox delivery." }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" } } } },
    "/api/mail/reply": { "post": { "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKeyRequired" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["conversationId", "messageId", "body"], "additionalProperties": false, "properties": { "conversationId": { "type": "string", "format": "uuid" }, "messageId": { "type": "string", "format": "uuid" }, "subject": { "type": "string", "maxLength": 200 }, "body": { "type": "string", "minLength": 1, "maxLength": 100000 }, "attachmentIds": { "type": "array", "maxItems": 10, "items": { "type": "string", "format": "uuid" } } } } } } }, "responses": { "201": { "description": "Reply queued for external SMTP delivery with In-Reply-To and References derived from the authorized incoming message." }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" } } } }
  },
  "components": {
    "parameters": {
...
      "IdempotencyKeyRequired": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 8, "maxLength": 128 } },
...
      "IfMatch": { "name": "If-Match", "in": "header", "required": false, "schema": { "type": "string", "pattern": "^\"revision-[1-9][0-9]*\"$", "example": "\"revision-1\"" } },
      "IfMatchRequired": { "name": "If-Match", "in": "header", "required": true, "schema": { "type": "string", "pattern": "^\"revision-[1-9][0-9]*\"$", "example": "\"revision-1\"" } },
...
      "DraftId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "ProviderTimestamp": { "name": "X-Provider-Timestamp", "in": "header", "required": true, "schema": { "type": "string" } },
      "ProviderSignature": { "name": "X-Provider-Signature", "in": "header", "required": true, "schema": { "type": "string" } }
    },
    "requestBodies": {
...
      "OtpRequest": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["phone", "purpose"], "properties": { "phone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "purpose": { "type": "string", "enum": ["signup", "login", "phone_change", "phone_change_old", "ivr_registration"] }, "channel": { "type": "string", "enum": ["sms", "ivr"] } } } } } },
      "OtpVerify": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OtpProof" } } } },
...
      "Draft": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "properties": { "subject": { "type": "string", "maxLength": 200 }, "body": { "type": "string", "maxLength": 100000 }, "to": { "type": "array", "maxItems": 50, "items": { "type": "string" } }, "cc": { "type": "array", "maxItems": 50, "items": { "type": "string" } } } } } } },
      "DraftSend": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "properties": { "conversationId": { "type": "string", "format": "uuid", "description": "Optional; omit to create or resolve a conversation from the saved To and CC recipients" }, "attachmentIds": { "type": "array", "maxItems": 50, "items": { "type": "string", "format": "uuid" } } } } } } },
...
      "MockProviderEvent": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["event"], "properties": { "event": { "type": "object", "required": ["id", "requestId", "status", "sequence", "occurredAt"], "properties": { "id": { "type": "string" }, "requestId": { "type": "string" }, "status": { "type": "string" }, "sequence": { "type": "integer" }, "occurredAt": { "type": "integer" } } } } } } } }
    },
    "schemas": {
      "OtpProof": { "type": "object", "required": ["phone", "challengeId", "purpose", "code"], "properties": { "phone": { "type": "string" }, "country": { "type": "string", "pattern": "^[A-Z]{2}$" }, "challengeId": { "type": "string", "format": "uuid" }, "purpose": { "type": "string" }, "code": { "type": "string", "pattern": "^[0-9]{6}$" } } },
      "OtpChallenge": { "type": "object", "required": ["challengeId", "purpose", "phone", "phoneE164", "country", "expiresAt", "channel"], "properties": { "challengeId": { "type": "string", "format": "uuid" }, "purpose": { "type": "string" }, "phone": { "type": "string", "description": "Canonical PhoneMail digit identity" }, "phoneE164": { "type": "string" }, "country": { "type": "string", "nullable": true }, "expiresAt": { "type": "string", "format": "date-time" }, "channel": { "type": "string", "enum": ["sms", "ivr"] } } },
...
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "requestId", "retryable"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" }, "fields": { "type": "object", "additionalProperties": true } } } } },
...
      "Draft": { "type": "object", "required": ["id", "revision"], "properties": { "id": { "type": "string", "format": "uuid" }, "subject": { "type": "string" }, "body": { "type": "string" }, "to": { "type": "array", "items": { "type": "string" } }, "cc": { "type": "array", "items": { "type": "string" } }, "revision": { "type": "integer", "minimum": 1 }, "updatedAt": { "type": "string", "format": "date-time" } } },
...
      "DraftPage": { "type": "object", "required": ["drafts"], "properties": { "drafts": { "type": "array", "items": { "$ref": "#/components/schemas/Draft" } } } },
...
      "SyncChangePage": { "type": "object", "required": ["changes", "cursor", "hasMore"], "properties": { "changes": { "type": "array", "items": { "type": "object", "required": ["cursor", "entity_type", "entity_id", "action", "payload"] } }, "cursor": { "type": "string", "pattern": "^(0|[1-9][0-9]*)$" }, "hasMore": { "type": "boolean" } } },
      "SyncSnapshotPage": { "type": "object", "required": ["snapshotId", "watermark", "records", "totalRecords", "hasMore", "nextCursor", "incrementalCursor"], "properties": { "snapshotId": { "type": "string", "format": "uuid" }, "watermark": { "type": "string", "pattern": "^(0|[1-9][0-9]*)$" }, "records": { "type": "array", "items": { "type": "object", "required": ["cursor", "entity_type", "entity_id", "payload"] } }, "totalRecords": { "type": "string", "pattern": "^(0|[1-9][0-9]*)$" }, "hasMore": { "type": "boolean" }, "nextCursor": { "type": "string", "nullable": true }, "incrementalCursor": { "type": "string", "nullable": true } } }
    },
    "responses": {
...
      "Error": { "description": "Stable error envelope with code, message, requestId and retryable", "headers": { "X-Request-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" }, "description": "Present for rate limits and retryable transient failures" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_session" },
      "refreshCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_refresh" }
    }
  }
}